backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-multipart
python-dotenv
httpx
requests
pytz
Pillow
apscheduler
